数値項目に小数点の規定値を設定できない問題の解消。

+++ Implem.Pleasanter/Models/Sites/SiteUtilities.cs
@@ -6259,7 +6259,7 @@ namespace Implem.Pleasanter.Models
                                                     controlId: "DefaultInput",
                                                     labelText: Displays.DefaultInput(context: context),
                                                     text: !column.DefaultInput.IsNullOrEmpty()
-                                                        ? column.DefaultInput.ToLong().ToString()
+                                                        ? column.DefaultInput.ToDecimal().ToString()
                                                         : string.Empty,
                                                     validateNumber: true,
                                                     _using: !column.Id_Ver
